splitting the code up
